Add location to content.
For now, if no location is set, new content defaults to current_user.location

rspec tests for models not expanded to use locations yet.
No way yet to edit location for content, for now it shows up on the location it is created from. (or on the users location as a fallback)