Adapting accessibility to:
 Location + find: N.A.
 Location + read: Without read access a location will not exist in retrieval.
Default for (group) all is read access on all locations.
 Location + publish: Allowing access to publish. Default being no such access.
 Content + find: "används av en location för att avgöra ifall location skall
visas i sökresultatet", i.e. retrieving locations with findable but not
necessarily readable content.
 Content + read: Without read access content will not exist in retrieval. A
location may thus be stated to contain a piece of content that will not appear
when retrieving content for that location, i.e. find but no read access for
the specific content.
 Content + publish: N.A.